Desktop-DTW-Python
We can prepare data sets to obtain speech features under various emotions, which can be as the source feature. The voice input by the user is used as the target feature. According to the similarity, we can construct the corresponding relationship between source features and target features to form feature pairs. Based on the pairs, we can find the most consistent emotion by using DTW algorithm.